+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## v1 rework with go-chi
+- 2024-07-27 : 
+    - add new front page to show budgets
+    - add a new date string field in financeTracker table
+    - remove old category func which used old fields and replace them with the rework
 - 2024-07-26 : 
     - change stats calculation to stop showing negative and positive value for the same category, now only show 1 total for each category
     - remove the autoreload on change in the category front page